+++ package.json
@@ -1,9 +1,26 @@
 {
-  "name": "my-app-frontend",
-  "version": "0.0.0",
+  "name": "@profcomff/ui-kit",
+  "version": "0.0.1",
   "type": "module",
+  "main": "./dist/ui-kit.umd.cjs",
+  "module": "./dist/ui-kit.es.js",
+  "types": "./dist/index.d.ts",
+  "styles": "./dist/ui-kit.css",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/ui-kit.es.js",
+      "require": "./dist/ui-kit.umd.cjs"
+    },
+    "./style.css": "./dist/ui-kit.css"
+  },
+  "files": [
+    "./dist"
+  ],
   "scripts": {
     "dev": "vite",
+    "type-check": "vue-tsc --noEmit",
+    "prepublishOnly": "pnpm build",
     "build": "vue-tsc && vite build",
     "docker-build": "vue-tsc && vite build --base=/ui",
     "preview": "vite preview",
@@ -12,64 +29,75 @@
     "lint:circular": "dpdm --exit-code circular:1 --no-tree --no-warning --progress false --transform ./src/main.ts",
     "format": "prettier src/ --write",
     "stylelint": "stylelint 'src/**/*.{vue,css}' --fix",
-    "check": "vue-tsc && pnpm format && pnpm lint && pnpm stylelint",
+    "check": "vue-tsc && pnpm format && pnpm lint",
     "check:hard": "vue-tsc && pnpm lint && pnpm lint:deadcode && pnpm lint:circular && pnpm format && pnpm stylelint",
     "test": "echo ok",
     "sb": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
-  "dependencies": {
+  "peerDependencies": {
     "@fontsource/roboto": "^5.2.10",
-    "storybook": "^10.1.0",
-    "vue": "^3.5.25"
+    "vue": "^3.5.0",
+    "vuetify": "^3.11.0"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "4.1.3",
-    "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.39.1",
+    "@fontsource/roboto": "^5.2.10",
     "@mapbox/stylelint-processor-arbitrary-tags": "^0.4.0",
     "@mdi/font": "^7.4.47",
-    "@profcomff/api-uilib": "2025.11.1-test",
-    "@storybook/addon-a11y": "10.1.0",
-    "@storybook/addon-docs": "10.1.0",
-    "@storybook/addon-vitest": "10.1.0",
-    "@storybook/vue3-vite": "10.1.0",
-    "@types/node": "^24.10.1",
-    "@typescript-eslint/eslint-plugin": "^8.48.0",
-    "@typescript-eslint/parser": "^8.48.0",
-    "@vitejs/plugin-vue": "^6.0.2",
-    "@vitest/browser": "^4.0.14",
-    "@vitest/coverage-v8": "^4.0.14",
-    "@vue/eslint-config-typescript": "^14.6.0",
-    "dpdm": "^3.14.0",
-    "eslint": "^9.39.1",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-prettier": "^5.5.4",
-    "eslint-plugin-storybook": "10.1.0",
-    "eslint-plugin-vue": "^10.6.1",
-    "knip": "^5.70.2",
-    "openapi-typescript": "^7.10.1",
-    "playwright": "^1.57.0",
-    "postcss-html": "^1.8.0",
-    "postcss-preset-env": "^10.4.0",
-    "prettier": "^3.6.2",
-    "sass": "^1.94.2",
-    "sass-loader": "^16.0.6",
-    "storybook-addon-pseudo-states": "^10.1.0",
-    "stylelint": "^16.26.0",
-    "stylelint-config-recommended": "^17.0.0",
+    "@profcomff/api-uilib": "2026.2.28",
+    "@storybook/addon-a11y": "10.3.0",
+    "@storybook/addon-docs": "10.3.0",
+    "@storybook/addon-vitest": "10.3.0",
+    "@storybook/vue3-vite": "10.3.0",
+    "@types/node": "^25.5.0",
+    "@typescript-eslint/parser": "^8.57.1",
+    "@vitejs/plugin-vue": "^6.0.5",
+    "@vue/eslint-config-typescript": "^14.7.0",
+    "dpdm": "^4.0.1",
+    "eslint": "^10.0.3",
+    "eslint-plugin-storybook": "10.3.0",
+    "eslint-plugin-vue": "^10.8.0",
+    "knip": "^5.88.1",
+    "postcss-preset-env": "^11.2.0",
+    "prettier": "^3.8.1",
+    "sass": "^1.98.0",
+    "storybook": "^10.3.0",
+    "storybook-addon-pseudo-states": "^10.3.0",
+    "stylelint": "^17.4.0",
     "stylelint-config-recommended-vue": "^1.6.1",
-    "stylelint-config-standard": "^39.0.1",
-    "stylelint-use-nesting": "^6.0.0",
+    "stylelint-config-standard": "^40.0.0",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.48.0",
+    "typescript-eslint": "^8.57.1",
     "unplugin-fonts": "^1.4.0",
-    "vite": "^7.2.4",
-    "vite-plugin-stylelint": "^6.0.2",
-    "vite-plugin-vuetify": "^2.1.2",
-    "vite-svg-loader": "^5.1.0",
-    "vitest": "^4.0.14",
-    "vue-tsc": "^3.1.5",
-    "vuetify": "^3.11.0"
-  }
+    "vite": "^7.3.1",
+    "vite-plugin-dts": "^4.5.4",
+    "vite-plugin-static-copy": "^3.3.0",
+    "vite-plugin-stylelint": "^6.1.0",
+    "vite-plugin-vuetify": "^2.1.3",
+    "vite-svg-loader": "^5.1.1",
+    "vue": "^3.5.30",
+    "vue-tsc": "^3.2.6",
+    "vuetify": "^3.12.3"
+  },
+  "description": "profcomff ui-kit based on vuetify 3 theme override and custom components",
+  "dependencies": {
+    "eslint-scope": "^9.1.2",
+    "eslint-visitor-keys": "^5.0.1",
+    "vue-eslint-parser": "^10.4.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/profcomff/ui-kit.git"
+  },
+  "keywords": [
+    "ui-kit",
+    "profcomff",
+    "vuetify3"
+  ],
+  "author": "batuevio",
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/profcomff/ui-kit/issues"
+  },
+  "homepage": "https://github.com/profcomff/ui-kit#readme"
 }
