+++ src/assets/styles.css
@@ -1,16 +1,16 @@
 .v-checkbox .v-selection-control {
-    gap: 4px;
+	gap: 4px;
 }
 
 .v-btn--icon.v-btn--density-compact {
-    width: calc(var(--v-btn-height) + -12px) !important;
-    height: calc(var(--v-btn-height) + -12px) !important;
+	width: calc(var(--v-btn-height) + -12px) !important;
+	height: calc(var(--v-btn-height) + -12px) !important;
 }
 
 .v-btn--icon .v-icon {
-    --v-icon-size-multiplier: calc(1);
+	--v-icon-size-multiplier: calc(1);
 }
 
 .v-icon--size-default {
-    font-size: calc(var(--v-icon-size-multiplier) * 1.5rem);
+	font-size: calc(var(--v-icon-size-multiplier) * 1.5rem);
 }
