--- src/pinia.ts
@@ -1,3 +1,0 @@
-import { createPinia } from 'pinia';
-
-export const pinia = createPinia();
